requirements.txt
Pillow
matplotlib
graphviz